applying api versioning to the project
