+++ src/api/v1/dto/EndpointResponse.dto.ts
