+++ src/api/v1/dto/hc/healthCheckResponse.dto.ts
