+++ src/api/v1/routes/todo.routes.ts
@@ -1,10 +1,10 @@
 import express, { Response } from "express";
-import authorize from "@middleware/authorize";
-import validator from "@middleware/validator";
-import validateResult from "@middleware/validationResults";
-import { userLimiter } from "@middleware/ratelimit";
+import authorize from "@/api/v1/middleware/authorize";
+import validator from "@/api/v1/middleware/validator";
+import validateResult from "@/api/v1/middleware/validationResults";
+import { userLimiter } from "@/api/v1/middleware/ratelimit";
 import todoService from "@services/todoService";
-import todoController from "@controllers/todoController";
+import todoController from "@/api/v1/controllers/todoController";
 import {
   ListTodosByOwnerRequest,
   ListTodoByOwnerRequest,
