+++ src/services/roleService.ts
@@ -9,8 +9,8 @@ import {
 } from "@/types/role.types";
 import Role from "@models/Role";
 import { paginate } from "@/services/paginationService";
-import { PaginatedResponseDTO } from "@/api/dto/pagination/pagination.dto";
-import { ErrorResponseDTO } from "@/api/dto/error/ErrorResponse.dto";
+import { PaginatedResponseDTO } from "@/api/v1/dto/pagination/pagination.dto";
+import { ErrorResponseDTO } from "@/api/v1/dto/error/ErrorResponse.dto";
 
 const listRoles = async function (
   params: ListRolesRequest
